Fixed listRoles jooq query

+++ application/src/main/java/net/furizon/backend/feature/roles/dto/ListedRoleResponse.java
@@ -18,5 +18,5 @@ public class ListedRoleResponse {
     private final long permanentUsersNumber;
     private final long temporaryUsersNumber;
 
-    private final long permissionsNumber;
+    private final int permissionsNumber;
 }
